+++ packages/dht/src/connection/webrtc/iceServerAsString.ts
@@ -1,4 +1,4 @@
-import { IceServer } from './WebrtcConnectorRpcLocal'
+import { IceServer } from './WebrtcConnector'
 
 export function iceServerAsString({ url, port, username, password, tcp }: IceServer): string {
     const [protocol, hostname] = url.split(':')
