+++ packages/dht/src/exports.ts
@@ -9,5 +9,5 @@ export { ConnectionManager, ConnectionLocker, PortRange, TlsCertificate } from '
 export { DhtRpcOptions } from './rpc-protocol/DhtRpcOptions'
 export { Remote } from './dht/contact/Remote'
 export { areEqualPeerDescriptors } from './helpers/peerIdFromPeerDescriptor'
-export { IceServer } from './connection/webrtc/WebrtcConnectorRpcLocal'
+export { IceServer } from './connection/webrtc/WebrtcConnector'
 export { DhtCallContext } from './rpc-protocol/DhtCallContext'
